pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["config", "data", "db", "bot", "scheduler", "utils"]

# Import restriction rules (enforce layer boundaries)
[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
]
markers = [
    "benchmark: Performance benchmark tests",
    "chaos: Chaos engineering tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/__init__.py",
    "*/conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 90

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "caktykbot"
version = "0.1.0"
description = "Automated stock trading bot for Indonesian market"
requires-python = ">=3.11"
dependencies = [
    "pymongo>=4.6.1",
    "yfinance>=0.2.36",
    "pandas>=2.2.0",
    "python-telegram-bot>=20.7",
    "APScheduler>=3.10.4",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "pytz>=2024.1",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "tabulate>=0.9.0",
]
